# docs/api/openapi/auth-service-openapi.json
{"openapi":"3.0.1","info":{"title":"Auth Service API","description":"Microservicio de autenticación JWT — InnovaTech DuocUC DSY1106-003V","contact":{"name":"Equipo InnovaTech","email":"dev@innovatech.cl"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8001","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Autenticacion","description":"Endpoints de login, registro, refresh y logout"}],"paths":{"/api/v1/auth/internal/users/{usersServiceId}":{"put":{"tags":["Autenticacion"],"summary":"Sincronizar credenciales desde users-service","operationId":"syncCredentials","parameters":[{"name":"usersServiceId","in":"path","description":"ID del usuario en users-service","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"requestBody":{"description":"Email, password y estado activo a sincronizar en auth-service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncCredentialsRequest"},"examples":{"syncCredentials":{"description":"syncCredentials","value":{"email":"codex.verifier@innovatech.cl","password":"Codex2026!","active":true}}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthUserResponse"}}}}}},"delete":{"tags":["Autenticacion"],"summary":"Eliminar credenciales vinculadas a un usuario de users-service","operationId":"deleteCredentials","parameters":[{"name":"usersServiceId","in":"path","description":"ID del usuario en users-service","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/register":{"post":{"tags":["Autenticacion"],"summary":"Crear credenciales para un usuario de users-service","operationId":"register","requestBody":{"description":"Datos de credenciales que seran vinculadas a un usuario existente en users-service","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"},"examples":{"registro":{"description":"registro","value":{"email":"codex.verifier@innovatech.cl","password":"Codex2026!","usersServiceId":3}}}}},"required":true},"responses":{"201":{"description":"Credenciales creadas","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserResponse"},"examples":{"registroExitoso":{"description":"registroExitoso","value":{"id":2,"email":"codex.verifier@innovatech.cl","usersServiceId":3,"activo":true,"bloqueado":false,"ultimoLogin":null,"createdAt":"2026-06-19T08:30:00"}}}}}},"409":{"description":"Email ya registrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthUserResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Autenticacion"],"summary":"Obtener nuevo access token usando el refresh token","operationId":"refresh","requestBody":{"description":"Refresh token emitido previamente en login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"},"examples":{"refresh":{"description":"refresh","value":{"refreshToken":"eyJhbGciOiJIUzI1NiJ9.refresh"}}}}},"required":true},"responses":{"401":{"description":"Refresh token invalido o expirado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"200":{"description":"Tokens renovados","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"},"examples":{"refreshExitoso":{"description":"refreshExitoso","value":{"accessToken":"eyJhbGciOiJIUzI1NiJ9.newAccess","refreshToken":"eyJhbGciOiJIUzI1NiJ9.newRefresh","tokenType":"Bearer","expiresIn":900,"userId":1,"email":"admin@innovatech.cl","permissions":["ROLE_ADMIN"]}}}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Autenticacion"],"summary":"Revocar refresh token y cerrar sesion","operationId":"logout","requestBody":{"description":"Refresh token a revocar","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"},"examples":{"logout":{"description":"logout","value":{"refreshToken":"eyJhbGciOiJIUzI1NiJ9.refresh"}}}}},"required":true},"responses":{"204":{"description":"Sesion cerrada y refresh token revocado"}},"security":[{"bearerAuth":[]}]}},"/api/v1/auth/login":{"post":{"tags":["Autenticacion"],"summary":"Login con email y contrasena y retorno de access plus refresh token","operationId":"login","requestBody":{"description":"Email y contrasena del usuario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"},"examples":{"loginAdmin":{"description":"loginAdmin","value":{"email":"admin@innovatech.cl","password":"Admin2024!"}}}}},"required":true},"responses":{"401":{"description":"Credenciales incorrectas","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"403":{"description":"Cuenta bloqueada o inactiva","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"200":{"description":"Login exitoso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"},"examples":{"loginExitoso":{"description":"loginExitoso","value":{"accessToken":"eyJhbGciOiJIUzI1NiJ9.access","refreshToken":"eyJhbGciOiJIUzI1NiJ9.refresh","tokenType":"Bearer","expiresIn":900,"userId":1,"email":"admin@innovatech.cl","permissions":["ROLE_ADMIN"]}}}}}}}}},"/api/v1/auth/change-password":{"post":{"tags":["Autenticacion"],"summary":"Cambiar contrasena del usuario autenticado","operationId":"changePassword","requestBody":{"description":"Contrasena actual y nueva contrasena segura","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"},"examples":{"changePassword":{"description":"changePassword","value":{"currentPassword":"Admin2024!","newPassword":"Admin2026!"}}}}},"required":true},"responses":{"204":{"description":"Contrasena actualizada"}},"security":[{"bearerAuth":[]}]}},"/api/v1/auth/validate":{"get":{"tags":["Autenticacion"],"summary":"Validar access token usado por el gateway y otros servicios","operationId":"validateToken","parameters":[{"name":"Authorization","in":"header","description":"Header Authorization con Bearer token","required":true,"schema":{"type":"string"},"example":"Bearer eyJhbGciOiJIUzI1NiJ9.access"}],"responses":{"200":{"description":"Token validado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateTokenResponse"},"examples":{"tokenValido":{"description":"tokenValido","value":{"valid":true,"email":"admin@innovatech.cl","usersServiceId":1,"permissions":["ROLE_ADMIN"]}}}}}}}}},"/api/v1/auth/users/{id}/exists":{"get":{"tags":["Autenticacion"],"summary":"Verificar si existe un authUser por ID para users-service","operationId":"existsById","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/v1/auth/test-hash":{"get":{"tags":["Autenticacion"],"summary":"Generar un hash de prueba para validaciones internas","operationId":"testHash","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"SyncCredentialsRequest":{"required":["active","email"],"type":"object","properties":{"email":{"type":"string"},"password":{"pattern":"^$|^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$","type":"string"},"active":{"type":"boolean"}}},"AuthUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"usersServiceId":{"type":"integer","format":"int64"},"activo":{"type":"boolean"},"bloqueado":{"type":"boolean"},"ultimoLogin":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"pattern":"^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$","type":"string"},"usersServiceId":{"type":"integer","format":"int64"}}},"RefreshTokenRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"email":{"type":"string"},"permissions":{"type":"array","items":{"type":"string"}}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"}}},"ChangePasswordRequest":{"required":["currentPassword","newPassword"],"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"pattern":"^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$","type":"string"}}},"ValidateTokenResponse":{"type":"object","properties":{"valid":{"type":"boolean"},"email":{"type":"string"},"usersServiceId":{"type":"integer","format":"int64"},"permissions":{"type":"array","items":{"type":"string"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}